+++ src/components/faqs/index.ts
@@ -1,0 +1,1 @@
+export { FaqsPage, type FaqsPageProps } from './faqs-page';